util module
