+++ wsmux/mux.go
@@ -4,6 +4,7 @@ import (
 	"time"
 
 	"github.com/gorilla/websocket"
+	"github.com/taskcluster/webhooktunnel/util"
 )
 
 // Config values for Session
@@ -11,7 +12,7 @@ type Config struct {
 	KeepAliveInterval    time.Duration // default 10 seconds
 	StreamAcceptDeadline time.Duration // default 30 seconds
 	RemoteCloseCallback  func()        // called when session is closed
-	Log                  Logger        // writes loggin messages. default is nilLogger
+	Log                  util.Logger   // writes loggin messages. default is nilLogger
 	StreamBufferSize     int           // default 1024 bytes
 }
